Adicionado Transaction

+++ DUDS/DUDS/Service/MovimentacaoPassivoService.cs
@@ -5,6 +5,7 @@ using DUDS.Service.SQL;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Data;
 using System.Threading.Tasks;
 
 namespace DUDS.Service
@@ -32,8 +33,22 @@ namespace DUDS.Service
 
             using (var connection = await SqlHelpers.ConnectionFactory.ConexaoAsync())
             {
-                string query = GenericSQLCommands.INSERT_COMMAND.Replace("TABELA", _tableName).Replace("CAMPOS", String.Join(",", _fieldsInsert)).Replace("VALORES", String.Join(",", _propertiesInsert));
-                return await connection.ExecuteAsync(query, item) > 0;
+                using (IDbTransaction transaction = connection.BeginTransaction())
+                {
+                    try
+                    {
+                        string query = GenericSQLCommands.INSERT_COMMAND.Replace("TABELA", _tableName).Replace("CAMPOS", String.Join(",", _fieldsInsert)).Replace("VALORES", String.Join(",", _propertiesInsert));
+                        var retorno = await connection.ExecuteAsync(sql: query, param: item, transaction: transaction);
+                        transaction.Commit();
+                        return retorno > 0;
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine(ex.Message);
+                        transaction.Rollback();
+                        return false;
+                    }
+                }
             }
         }
 
@@ -72,12 +87,25 @@ namespace DUDS.Service
         public async Task<bool> DeleteByDataRefAsync(DateTime dataMovimentacao)
         {
             List<OrdemPassivoModel> result = await GetByDataEntradaAsync(dataMovimentacao: dataMovimentacao) as List<OrdemPassivoModel>;
-
+            if (result == null) return false;
             using (var connection = await SqlHelpers.ConnectionFactory.ConexaoAsync())
             {
-                const string query = "DELETE FROM tbl_movimentacao_passivo WHERE data_movimentacao = @data_movimentacao";
-                int rowsAffected = await connection.ExecuteAsync(query, new { data_movimentacao = dataMovimentacao });
-                return rowsAffected > 0 && rowsAffected == result.Count;
+                using (IDbTransaction transaction = connection.BeginTransaction())
+                {
+                    try
+                    {
+                        const string query = "DELETE FROM tbl_movimentacao_passivo WHERE data_movimentacao = @data_movimentacao";
+                        int rowsAffected = await connection.ExecuteAsync(sql: query, param: new { data_movimentacao = dataMovimentacao }, transaction: transaction);
+                        transaction.Commit();
+                        return rowsAffected > 0 && rowsAffected == result.Count;
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine(ex.Message);
+                        transaction.Rollback();
+                        return false;
+                    }
+                }
             }
         }
 
@@ -130,14 +158,28 @@ namespace DUDS.Service
 
             using (var connection = await SqlHelpers.ConnectionFactory.ConexaoAsync())
             {
-                string query = GenericSQLCommands.UPDATE_COMMAND.Replace("TABELA", _tableName);
-                List<string> str = new List<string>();
-                for (int i = 0; i < _propertiesUpdate.Count; i++)
+                using (IDbTransaction transaction = connection.BeginTransaction())
                 {
-                    str.Add(_fieldsUpdate[i] + " = " + _propertiesUpdate[i]);
+                    try
+                    {
+                        string query = GenericSQLCommands.UPDATE_COMMAND.Replace("TABELA", _tableName);
+                        List<string> str = new List<string>();
+                        for (int i = 0; i < _propertiesUpdate.Count; i++)
+                        {
+                            str.Add(_fieldsUpdate[i] + " = " + _propertiesUpdate[i]);
+                        }
+                        query = query.Replace("VALORES", String.Join(",", str));
+                        var retorno = await connection.ExecuteAsync(sql: query, param: item, transaction: transaction);
+                        transaction.Commit();
+                        return retorno > 0;
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine(ex.Message);
+                        transaction.Rollback();
+                        return false;
+                    }
                 }
-                query = query.Replace("VALORES", String.Join(",", str));
-                return await connection.ExecuteAsync(query, item) > 0;
             }
         }
     }
